fix(EXC-1811): Disable `anyhow` backtraces in sandbox
EXC-1811

As of version 1.0.77, the `anyhow` crate will capture a backtrace any
time an error is generated (if `RUST_BACKTRACE` is set). This triggers
some SELinux denials in the sandbox because libunwind reads and writes
to a pipe when capturing the backtrace. We can disable the capturing in
the sandbox by setting the `RUST_LIB_BACKTRACE` env variable since we
don't use these backtraces anyway.