feat(IC-1579): Spawn neuron(s) TLA instrumentation (#2944)
This adds the TLA instrumentation for spawn_neuron/spawn_neurons.

Also extends the instrumentation framework to deal with non-async
methods (as `spawn_neuron` isn't async), fixes a bug in the
instrumentation (where a failed check would panic, leading to the
starvation of the channel listener), and adds state diffs to more easily
debug divergences from the model.