feat(ckbtc): add timestamps to suspended UTXOs (#2939)
Follow-up on #2674 to ensure that suspended UTXOs are re-evaluated at
most once a day.

Remarks:

1. The timestamps of when a suspended UTXO was checked the last time are
only stored on the heap and will not be preserved during canister
upgrades. This is deemed acceptable since the amount of suspended UTXOs
is typically fairly small (currently 68) and this avoids having to store
an additional event in stable memory every time a suspended UTXO is
re-evaluated.
2. Suspended UTXOs that cannot yet be re-evaluated are added to
`UpdateBalanceError::NoNewUtxos` in case there are currently no UTXOs to
process.