fix: Await all calls at once in xnet_test_canister
When xnet_test_canister was migrated from dfn_core to ic_cdk, it retained the structure of the earlier, callback-based fanout() method. Which meant that it was now awaiting the calls to eacu subnet separately. In the xnet_slo_120_subnets_staging_test, this lead to extended ramp-up and and ranp-down periods, which caused the test to time out.

Instead, create futures for calls to all subnets first; and only then await them all.