fix(nns): Change the default voting power refreshed timestamp from Nov 15 to Sep 1. (#2972)
Originally, we decided to grandfather existing neurons so that they
would be considered refreshed on Nov 15 (more precisely, [1731628801
seconds after UNIX epoch][used]), but later, we changed our minds, and
said that the clock should start on Sep 1 (more precisely, 1725148800).
We changed the value of the constant, but many neurons picked up the old
value. This sets those neurons to the new value.

[used]:
https://sourcegraph.com/github.com/dfinity/ic@1956e438af82a5b4aa9713bcbbe385684bf0704f/-/blob/rs/nns/governance/src/lib.rs?L189

# References

This is part of the "periodic confirmation" feature that was recently
approved in proposal [132411][prop].

[prop]: https://dashboard.internetcomputer.org/proposal/132411

Closes [NNS1-3488][closes].

[closes]: https://dfinity.atlassian.net/browse/NNS1-3488

[NNS1-3488]: https://dfinity.atlassian.net/browse/NNS1-3488