fix: Await all calls at once in xnet_test_canister's fanout() (#3008)
When `xnet_test_canister` was migrated from `dfn_core` to `ic_cdk`, it
retained the code structure of the earlier, callback-based `fanout()`
function. Which unfortunately meant that it was now awaiting the calls
to each subnet separately. In `xnet_slo_120_subnets_staging_test`, this
lead to extended ramp-up and and ramp-down periods, which caused the
test to time out.

Instead, create futures for calls to all subnets first; and only then
await them all.