fix(IDX): override mainnet_revisions.json
The replica release `2d8611eb4efa8e69c4dd567546c1c353a545e0a6` was built with Ubuntu-24.04 which means that it expects a newer version of libc to be available in the environment where its binaries run. After deploying this release we had to downgrade to Ubuntu 20.04 again. This means that the binaries from the latest replica release don't work in our tests since the latest libc is not available there. This effects the `//rs/tests/consensus/backup:backup_manager_upgrade_test_colocate` for example. We therefor temporarily override the mainnet revision to the commit just before the Ubuntu 24.04 upgrade.