feat(nns): Periodic confirmation metrics. (#3070)
Most of the changes here are for breaking out voting power into deciding
vs. potential, and copying those around, because of duplicate types.

Another metric in this PR: When the current prune following cycle began.

# References

This is part of the "periodic confirmation" feature that was recently
approved in proposal [132411][prop].

[prop]: https://dashboard.internetcomputer.org/proposal/132411

[Next PR 👉][next]

[next]: https://github.com/dfinity/ic/pull/3090