feat(nns): Count stale and expired voting power neurons. (#3090)
Here,

* stale = time since voting power was refreshed is between strictly 6
and 7 months (or whatever VotingPowerEconomics dictates).
* expired = time since voting power was refreshed is >= 7 months (again,
per VotingPowerEconomics).

I think the main thing we really care about here is the number of such
neurons in those two (sub)categories. However, it is not much harder to
gather additional statistics (thanks to existing code for compiling
statistics for arbitrary neuron (sub)categories). Therefore, a whole
bunch of other stats about these two subcategories of neurons is also
exported.

FWIW, if you strip away the highly repetitive code (and don't count test
code), this change is actually pretty small.

# References

This is part of the "periodic confirmation" feature that was recently
approved in proposal [132411][prop].

[prop]: https://dashboard.internetcomputer.org/proposal/132411

Closes [NNS1-3418].

[NNS1-3418]: https://dfinity.atlassian.net/browse/NNS1-3418

[👈 Previous PR][prev]

[prev]: https://github.com/dfinity/ic/pull/3070