refactor(tests): `await_with_timeout` should take a range, not just a timeout (#3096)
The issue is that there are some cases where we know that
`await_with_timeout` will have to wait a certain minimum amount of time.
By taking a range, we can have the caller pre-emptively tell us the
minimum amount of time they think we'll have to wait. Then we can
advance that amount of time straight away