feat: [EXC-1753] Add mint_cycles128 API (#2589)
This PR adds a new system API method `ic0_mint_cycles128` which, like
its predecessor `ic0_mint_cycles`, is not usable by any canister except
the CMC. It is therefore not necessary to add it to the interface spec
or the SDKs. It also means that this system call does not need an
instruction cost adjustment.

The method's signature deviates from the predecessor because 128 bit
values cannot be handled by WASM natively. Therefore, the argument is
split into two `u64`s, representing the high and low bits. The amount of
actually minted cycles cannot be returned as a WASM return value either.
Instead, the result is written to a caller-specified location in memory,
where it occupies the next 16 bytes.

---------

Co-authored-by: Michael Weigelt <michael.weigelt@dfinity.com>
Co-authored-by: mraszyk <31483726+mraszyk@users.noreply.github.com>