feat: CON-1418 CON-1421 Implement a stubbed VetKd client as part of the IDKG signer (#2936)
This PR implements the vetKD protocol by extending the existing
`ThresholdSigner` component that is already driving the tSchnorr and
tECDSA protocols. Conceptually, these protocols are very similar, as
they all consist of a single "share" type, which is broadcasted,
verified and combined. Therefore, combining the implementation as part
of the same component avoids having to implement an entirely new client
and artifact pool, which would only be used by the VetKd protocol.

One downside of this approach is the pointless transition of
`VetKdContext` -> `ThresholdSigInputsRef` -> `ThresholdSigInputs`, as
there are no "pre-signature refs" that would need to be resolved in
VetKd. However, a future improvement envisions storing pre-signatures in
the replicated state as opposed to the blockchain. At this point
`ThresholdSigInputsRef` would disappear entirely anyway.

Another downside is that VetKd will share the same thread as IDKG and
ECDSA/Schnorr signature generation. In a future iteration, we can
extract a new client (thread) that is used exclusively for VetKd, ECDSA
and Schnorr (separate from IDKG key and pre-signature generation).

For now, the client doesn't yet call the corresponding crypto endpoints
to create/verify/combine VetKd encrypted key shares, as these endpoints
aren't implemented yet. Once they are, additional unit tests exercising
the new protocol will be added as well.