feat(BOUN-1308): build and publish anonymization canister (#3106)
This adds a github workflow to manually trigger a build/release of the
anonymization canister. We plan to add a similar workflow for another
new rate-limiting canister.

I tested this on a small temporary repo but without the bazel portion,
so am not entirely sure if this would require another runner or also
installing ic-wasm? cc @basvandijk

---------

Co-authored-by: Bas van Dijk <bas@dfinity.org>