perf(nns): Add benchmarks for neuron data validator (#3102)
# Why

Neuron data validator goes through batches of neurons in order to make
sure neurons and its indexes are in a consistent state, and such
operations will get more expensive after neurons are migrated to stable
structures. Adding benchmarks makes it easier to optimize the
performance, and help prevent regressions.

# What

Add benchmarks for neuron data validator for 2 cases - whether the
neuron migration to stable structures is enabled or not.