feat(EXC-1780): Pass `SerializedModule` via file descriptor (#2850)
EXC-1780

Add support in the sandbox IPC to pass `SerializedModules` from the
`CompilationCache` via a file descriptor instead of serializing the
data. This PR doesn't actually enable the feature because the cache will
still be storing in "memory" mode instead of "disk".

Follow up PRs will limit the cache size on disk and use the new Wasmtime
option do directly map the file before the feature is enabled.