chore: EXC-1822: Increase max number of sandboxes to 7k (#3164)
In light of the slight performance degradations we have noticed on some
of the Yral subnets with slow FR which recovers after a few minutes, the
execution team would like to propose to upgrade the affected subnets to
include this patch early next week. Although the issue is certainly not
critical, it creates some noise and we believe that it should be better
to allow everyone to have a page-less holiday period.

The change should be very safe as it basically increases the cache of
sandboxes we keep around which is possible since a few other
improvements that have already been merged a couple weeks ago. It should
allow us to absorb better these cases where we’ve got some extra load by
canisters that have not been executed for a while before we max out the
sandbox cache.

The hope would be to elect a new replica version based on the latest
deployed RC, and then deploy it to the potentially affected subnets
early next week.