fix(ICP_Ledger): FI-1590: Change length type in GetBlocksArgs (#3092)
Change the type of the `length` field in the ledger `GetBlocksArgs`
struct from `usize` to `u64`. The field in the candid interface is of
type `Nat64`, but if a value larger than what fit into a `u32` was
passed in as the `length` in a `query_blocks` query, this would lead to
a `Fail to decode argument 0 from table0 to record { start : nat64;
length : nat64 }` error.