fix(nns): Fix benchmark by simulating proper behavior in canbench-rs build (#3160)
In order to get approximately the same number of read/writes in stable
for the VotingStateMachine, it is required to count accurately. Canbench
doesn't allow queries, which means the instruction_counter never gives a
different result, which means the benchmark thinks it runs out of cycles
on every iteration, and thus saves and restores the
ProposalVotingStateMachine over and over, leading to an inaccurate
benchmark compared to production.

This adds code to approximately simulate behavior in canbench-rs build.