feat(recovery): Avoid rsync on state uploads during local recoveries (#2948)
The purpose of this change is to avoid a costly data transfer of state,
in cases where we're performing recoveries directly on the admin node.
The bulk of the change is in `steps.rs`. For local recoveries, we don't
have to involve rsync or ssh, or copy the latest checkpoint.

From the operator's PoV, the only thing that changed was that instead of
being asked for the upload node's ip address, they receive the following
prompt:

> (Optional) Are you performing a local recovery directly on the node,
or a remote recovery? [local/\<ipv6\>]