fix: Record 'dropped' status for silently dropped best-effort responses. (#3162)
The assumption that `push_input()` returning `Ok(())` implies that the
message was successfully inducted is correct for guaranteed responses
(otherwise it's a critical error) but not for best-effort responses
since those may be dropped silently. It is therefore necessary to
distinguish between these two cases by returning `Ok(true/false)`
instead to record successful induction in the metrics only for actual
inductions and 'dropped' otherwise.