fix: Add exceptions for edge cases for best effort responses. (#3155)
For best effort responses, it is expected that sometimes a second
response arrives late. These cases are filtered by checking against the
callback when inducting into the replicated state, but the canister
could have been stopped or even removed in the mean time in which case
we don't even make it to this check. Therefore these cases must be
filtered elsewhere since they are expected and not critical errors.