fix: including dashboards into colocated uvm's when running tests (#3219)
Previously the UVM's would fail to copy the dashboards over to the farm
testnet because they were not copied from local to colocated UVM.
This PR adds the missing functionality. 

Also fixes the `file_name` formatting. It removes an additional `.json`
added to the dashboard name which already includes the extensions.