feat: Add wasm_memory_threshold to ProposeToUpdateCanisterSettingsCmd (#2864)
To support `on_low_wasm_memory` hook we are adding
`wasm_memory_threshold` to `NNS` so it allows users to update its value
using `NNS` proposal. We are as well adding `wasm_memory_threshold` to
canister settings to canisters in `IC` repo.

The `wasm_memory_threshold` will be added to the interface specification
in https://github.com/dfinity/portal/pull/3761

---------

Co-authored-by: Andre Popovitch <andre@popovit.ch>
Co-authored-by: Arshavir Ter-Gabrielyan <arshavir.ter.gabrielyan@dfinity.org>