feat: Adding syncing of NNS released wasm versions (#3148)
This PR refactors the script used for updating mainnet revisions in
testnets and adds the ability for it to be used more robustly while
preserving the existing features.

On top of that, this PR extends the `update-mainnet-revisions` workflow
to add one more job that will be run in parallel and will update the
`mainnet-canisters.json` file.

Example of new successful run:
https://github.com/dfinity/ic/actions/runs/12297897124
Example of created PR: https://github.com/dfinity/ic/pull/3198