chore(ckbtc): add optional timestamp to minter events (#3157)
[(XC-238)](https://dfinity.atlassian.net/browse/XC-238) ckBTC events, in
contrast with ckETH events, do not have a timestamp. This makes it
almost impossible to determine when something happened. To address this,
this PR adds a timestamp to ckBTC minter events.

In order for this change to be backwards compatible, the timestamp is
optional and events logged prior to the introduction of timestamps have
a `None` value for this field. The deserialization of both legacy and
new events is supported, however there is a breaking change to the
Candid interface of the `get_events` endpoint of the ckBTC minter (hence
the `CI_OVERRIDE_DIDC_CHECK` label) due to the new structure of the
`Event` type (i.e. addition of a timestamp field). This is acceptable
since it is a debug endpoint.