feat(crypto): CRP-2649 add vault method to create vetKD key share (#2989)
Adds a trait `VetKdCspVault` with a method
`create_encrypted_vetkd_key_share` and implements it for the local and
the remote crypto vault.

Also adapts the type of `VetKdEncryptedKeyShare::encrypted_key_share`
from `Vec<u8>` to a new `VetKdEncryptedKeyShareContent` type (which is
returned by `VetKdCspVault::create_encrypted_vetkd_key_share`) so that
we can implement the `DomainSeparator` trait for it, so that we can
later use the default infrastructure for signing this
`VetKdEncryptedKeyShareContent`.