feat: [NODE-1549] Disable start/stop logs for timers (#3244)
Brings back the old branch from `dfinity-lab/public/ic!18736`, now that
we have a supported systemd.

For timer units, we set the default log level above the max filter, to
clear the noisy "Starting" and "Finished" logs. Using a max of 1 gives
us two log levels to work with, but we should prefer 1 as 0 broadcasts
to the system. Log level can be set with a `<1>` prefix on any line, but
these units tend to only write metrics, anyway.