test(cketh): end-to-end test with `foundry` (#3014)
[XC-256](https://dfinity.atlassian.net/browse/XC-256): End-to-end system
test for the whole ckETH/ckERC20 setup with a local Ethereum deployment
environment based on [foundry](https://book.getfoundry.sh/), which is
made of the following components:

1. ckETH minter canister
2. ckETH ledger canister
3. Ledger suite orchestrator to manage ckERC20 ledger suites
4. Deployment of helper smart contracts via `foundry` (`anvil`).

The current test focusses on the interaction with the helper smart
contracts whose addresses are retrieved from the minter information to
ensure that:

1. the smart contract ABI does not change (happened once in production,
see proposal
[134344](https://dashboard.internetcomputer.org/proposal/134344)).
2. User-deposited funds are moved to the minter's address.

**Out of scope**: the minter currently doesn't retrieve the minting
events that would require the EVM RPC canister and is deferred to
[XC-258](https://dfinity.atlassian.net/browse/XC-258). The reason is
that it's currently not possible to intercept in a system test calls
made by the EVM RPC canister:
1. The caller of the EVM RPC canister, which is the ckETH minter, uses a
hard-coded 3-out-of-4 strategy, where the choice of providers is up to
the EVM RPC canister.
    3. Providers and their URLs are hard-coded in the EVM RPC canister.
4. HTTPs outcalls made by canisters do not go through boundary nodes and
the DNS servers to do hostnames resolution are hard-coded in the
replica.

[XC-258]:
https://dfinity.atlassian.net/browse/XC-258?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

[XC-256]:
https://dfinity.atlassian.net/browse/XC-256?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ