feat(recovery): [CON-1425] Allow skipping nodes during certification downloads (#3188)
This PR adds a way to manually skip nodes during certification downloads
when rsync fails.

Background: During recoveries, we should download and merge the
certification pools of as many nodes on the subnet as possible.
Sometimes there are connection issues to individual nodes, and the tool
gets stuck trying to download the node’s pool. In those cases we may
want to skip the problematic nodes.