feat: Implement MutableIntMap
MutableIntMap is an internally mutable variant of IntMap. The underlying tree is still a persistent data structure, so different copies of the same MutableIntMap can be modified independently. And the performance is similar, since it's based on the same internals.

It is intended as a drop-in replacement for BTreeMap (hence the internal mutability as opposed to a purely functional interface). For that reason, it also implements remove(), split_off() and extend(); as well as a couple of syntactic sugar methods (e.g. keys() and values()).

The main performance difference between IntMap and MutableIntMap comes from (i) having insert() return the previous value; and (ii) keeping track of / recalculating the size of the map, so that len() runs in constant time.