perf: Put canister queues behind Arcs
This makes unmutated canister queues basically free to clone, which is important during certification and, more importantly, before and after every message execution (when the SystemState is cloned in order to allow for rollbacks). After the changes in #3241 and #3304, this accounts for about half the execution time; droppinbg to well under 10% with this change.