fix(nns): Fix for a particular locked neuron (#3311)
A neuron was locked, which logged `Inflight commands: {
17912780790050115461: NeuronInFlightCommand { timestamp: 1728911670,
command: Some( Spawn( NeuronId { id: 17912780790050115461, }, ), ), },
}`

We identified the account as:
https://dashboard.internetcomputer.org/account/9765cc69bd6580023bee8fa8280fb630977b9f850da49d73cddcb99512272abb
And the neuron can be seen here:
https://dashboard.internetcomputer.org/neuron/17912780790050115461

As can be seen, no ICP was minted, but the neuron's cached stake was set
and maturity was cleared. The fix is to mint the ICP and unlock the
neuron.

Fix to prevent this from happening in the future is here:
https://github.com/dfinity/ic/pull/3226