feat(PocketIC): optionally check caller when retrieving ingress status (#3341)
This PR extends the PocketIC functionality of retrieving ingress status
with an optional caller: if provided and if a corresponding read state
request for the status of the same update call signed by that specified
caller was rejected because the update call was submitted by a different
caller, then an error is returned. To distinguish between the ingress
status not being available and forbidden due to the new error condition
introduced by this PR, a new enumeration `IngressStatusResult` is
introduced in the PocketIC library.