feat: ICRC-ledger: FI-1439: Implement V4 for ICRC ledger - migrate balances to stable structures (#2901)
This version of the ledger migrates the balances stored by the ledger
from heap to stable memory stable structures. This allows the ledger to
store more data (heap memory is limited to 4GB) and reduces the number
of instructions used during the upgrades - there is no need to serialize
and deserialize the balances stored in heap memory. Since we are no
longer limited by the heap size, the logic for trimming balances and
related init/upgrade arguments (`maximum_number_of_accounts`,
`accounts_overflow_trim_quantity`) were removed. The migration to stable
structures is performed during the upgrade and, if 200B instruction
limit is reached, using timer invocations after the upgrade. If timer
invocations are used by the migration, the ledger is unavailable for
transfers and approvals until migration is finished. The migration
status can be checked by checking the value of the
`ledger_stable_upgrade_migration_steps` metric and calling the
`is_ledger_ready` endpoint. The metric will increase at roughly 1s
intervals until the migration is finished. The canister logs will also
contain some information about the number of migrated allowances and
balances (e.g. "Migration partially done. Scheduling the next part.
Number of elements migrated: allowances: 0 expirations: 0 balances:
7657..."). If migration is not progressing and the ledger is not ready,
the ledger can still be downgraded to the previous version. After
migration is finished and `is_ledger_ready` returns true, it is no
longer possible to downgrade the ledger to the previous version .

---------

Co-authored-by: Mathias Björkqvist <mathias.bjorkqvist@dfinity.org>