chore: Update ic-agent and ic-agent's reverse dependencies to v0.39 (#3282)
I created this PR by following these instructions from @nathanosdev:

> The following crates should use version 3.0.2:
> 1. `ic-response-verification`
> 2. `ic-certification`
> 3. `ic-certificate-verification`
> 4. `ic-http-certification`
> 5. `ic-cbor`
> 
> I recommend just setting them to `3` in `Cargo.toml` and `^3.0.2` in
`external_crates.bzl` so they can be updated more easily in the future.
That's how it's setup in `ic-http-gateway` too.
> And `ic-http-gateway` can be set to version `0.1.0`. The git tag
reference can be removed now.

I also updated `ic-agent` to 0.39 because that is the version of
`ic-agent` that `ic-http-gateway` v0.1.0 depends on. In doing so, I also
updated `ic-utils` and `ic-transport-types`. I also moved the
`ic-http-gateway` dependency to the workspace-level `Cargo.toml`.

It seems like there were a number of breaking changes in these crates,
so I also went through and fixed all the code that became broken.

---------

Co-authored-by: IDX GitLab Automation <idx@dfinity.org>