chore: CON-1434 Increment metric in case of `NiDkgTag` mismatches (#3329)
With https://github.com/dfinity/ic/pull/3039 we will start to serialize
maps of the form `(nidkg_tag, nidkg_transcript)` as a vector of
`nidkg_transcript`s. This is fine since the NiDkg tag is also part of
the NiDkg transcript. More generally, the change makes the assumption
`nidkg_tag == nidkg_transcript.tag` explicit.

As an extra pre-caution, this PR adds a log and metric to increase our
confidence that this assumption already holds today (and thus that the
change in https://github.com/dfinity/ic/pull/3039 is safe to roll out).