fix: Include reservation cycles in error reported during taking a canister snapshot (#3331)
When taking a canister snapshot, the balance of the canister is checked
to confirm that it has enough cycles to cover its freezing threshold as
well as any reservation of cycles needed in case the subnet's capacity
is above the storage reservation threshold. However, in the error
reported in case the canister does not have sufficient balance, only the
freezing threshold was included, thus resulting in cases where an
incorrect required amount would be reported to the user. This PR fixes
the issue and adds some tests to ensure that there are no regressions in
the future.