refactor: Rename field in InsufficientCyclesInMemoryGrow error (#3333)
The field `threshold` in `InsufficientCyclesInMemoryGrow` error is not
quite accurate. Some times it's not only the freezing threshold that
might need to be taken into account (assuming the naming comes from
that) but also other sources of consuming cycles, like for example the
storage reservation cycles. Rename the field to `required` which also
matches with the naming of other similar errors that deal with out of
cycles cases.