feat(recovery): [CON-1407] Add CLI option to use existing binaries for local recoveries (#3301)
This PR adds `ic-admin` to the guest OS & adds a CLI argument to make
the recovery tool use the local binaries from `/opt/ic/bin`.

With this, we can kick off a non-interactive, local subnet recovery via
ssh, by calling the recovery tool with the right arguments. With this
change, we should be able to create a system test for local recoveries.


To confirm that this works, I performed a non-interactive recovery on a
testnet with the following arguments:

```
/opt/ic/bin/ic-recovery \
	--nns-url http://[2602:fb2b:110:10:5025:8bff:fe14:daee]:8080 \
	--test \
	--skip-prompts \
	--use-local-binaries \
	app-subnet-recovery \
	--subnet-id c3es6-ogziq-a5elm-47c4o-vkins-7kiu4-ekgun-oaeo2-ojd2m-j26wh-eae \
	--pub-key "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIFx8H9eVvkxl1n7crh/uUzqH03m8Qwlo2k5EE3MJa4Bs adrian.alic@dfinity.org" \
	--download-node 2602:fb2b:110:10:506a:22ff:fe1c:ef69 \
	--upload-method local \
	--keep-downloaded-state false
```

The subnet recovered successfully without user input.