feat(governance): Enable periodic confirmation. (#3371)
This is done by flipping a couple of flags.

This is the last step in the release of "periodic confirmation". There
is still some post-launch follow up work.

All the usual consequences of this being a flag flip apply. In
particular, if we need to roll back, it will be fairly easy and low
risk.

This causes a performance regression for metrics. [A ticket][regression]
has been opened to follow up on that. For now, this has been deemed
acceptable, because metrics are calculated in the background, and is
performed infrequently.

[regression]: https://dfinity.atlassian.net/browse/NNS1-3535

# References

This is part of the "periodic confirmation" feature that was recently
approved in proposal [132411][prop].

[prop]: https://dashboard.internetcomputer.org/proposal/132411

Closes [NNS1-3444].

[NNS1-3444]: https://dfinity.atlassian.net/browse/NNS1-3444