feat(BOUN-1333, BOUN-1334): ratelimit: add scaling, change allow rule, change shedding latencies (#3379)
* Adds optional autoscaling of rules to the ratelimiter (CLI arg
`--rate-limit-generic-autoscale`). If enabled it will divide the limit
by the number of API BNs in the registry and recalculate whenever the
number changes (when the new snapshot is published)
* Remove implicit allow rule for the ratelimit canister, instead just
bypass the ratelimiting for the requests that originate from localhost.
This allows to create ratelimit rules for the canister itself
* Change HTTP code from `429` to `403` when rejecting requests by the
rule that has `limit: block`
* Bump load shedding latencies in `ic-gateway` to fix the false
positives that we're getting occasionally

---------

Co-authored-by: IDX GitLab Automation <idx@dfinity.org>