chore: state-tool copy commands
This commit introduces the "copy" subcommand to the state-tool copying
checkpoints from one place to another.

This command can both be used to backup a state as well as importing a
state to an existing directory. It handles both the checkpoint files
itself as well as the accompanying metadata such as manifests.

The new subcommand is more general than the existing "import"
subcommand, as the old one only allows for imports and does not handle
metadata. This commit does however not remove "import" in case somebody
relies on it, and simplifies its internals a bit instead.