test: [EXC-1819] Fix flakiness of state_manager_integration_tests (random_canister_input_lsmt) (#3200)
The test is flaky since freeing of sandbox threads, commit `4a622c0`
(`RUN-1014`), which was modified at `EXC-1681`(`82c76c1`), but not
resolved. On top of that `a1e516f` introduced more flakiness.
One reason is `StateManager` outliving the `StateMachine` because
something holds the `Arc` containing it, yet we remove & recreate
tempdir using the same name, ending up with two `StateManagers`
accessing the same dir. This is addressed by reordering fields of
`StateMachine` to ensure we destruct `StateManager` before the StateDir
at `StateMachine::drop` and waiting for the `StateManager` to drop in
`StateMachine::into_components`.
Another issue is communication with sandbox, it was solved by using fork
in `ProptestConfig`.