test(consensus): modify the test cases in ingress payload serialization/deserialization benchmarks (#3447)
Currently we only measure how much it takes to deserialize/serialize
ingress payloads with 2000/4000/6000/8000 messages each of size 1KB.
With the changes in this PR we will instead measure the following test
cases:
1. 1000 messages, 4KB each
2. 2000 messages, 4KB each
3. 1 message, 4MB
4. 1 message, 8MB

which resembles more closely the current limits on the ingress payload,
from both number of messages and message sizes extremes.

Also, in the deserialization benchmark we moved out the `clone()` as it
affects the measurements significantly.