refactor(consensus): simplify `IngressPayload` implementation (#3444)
Currently the payload is one big bytes buffer which contains all ingress
messages in a packed representation. This makes it a big hard to work
with.
In this PR we simplify the structure by replacing the buffer with a map
from `IngressMessageId`s to a _serialized_ ingress messages.
The following two important properties are preserved:
1. The individual ingress messages deserialization is delayed until it's
actually needed
2. We preserve the original byte representation of the ingress messages

I've ran multiple benchmarks and didn't notice a _big_ change:
1. `consensus-performance` system test showed the same throughput /
block rates with both implementations
2. the [serialization/deserialization
](https://github.com/dfinity/ic/blob/master/rs/consensus/benches/validate_payload.rs#L374-L404)
of ingress payload gets about 2x slower, but it's still in sub 1ms
territory

---------

Co-authored-by: Leon Tan <leon.tan@dfinity.org>